Kernel Density Estimation
Source: R/binned.kernel.est.R

This function implements an interface to the kernel density estimation functions provided by the KernSmooth package.

binned.kernel.est(data, kernel = "normal", bandwidth = NULL,
  canonical = FALSE, scalest = "minim", level = 2L, gridsize = 401L,
  range.data = range(data), truncate = TRUE)

Arguments

data
a numeric vector containing the sample on which the kernel density estimate is to be constructed.

kernel
character string specifying the smoothing kernel

bandwidth
the kernel bandwidth smoothing parameter.

canonical
a logical value indicating whether canonically scaled kernels should be used

scalest
estimate of scale.

"stdev" - standard deviation is used.

"iqr" - inter-quartile range divided by 1.349 is used.

"minim" - minimum of "stdev" and "iqr" is used.

level
number of levels of functional estimation used in the plug-in rule.

gridsize
the number of equally-spaced points over which binning is performed to obtain kernel functional approximation.

range.data
vector containing the minimum and maximum values of data at which to compute the estimate. The default is the minimum and maximum data values.

truncate
logical value indicating whether data with x values outside the range specified by range.data should be ignored.

References

Matt Wand (2015). KernSmooth: Functions for Kernel Smoothing Supporting Wand & Jones (1995). R package version 2.23-14.

Henry Deng and Hadley Wickham (2011). Density estimation in R. http://vita.had.co.nz/papers/density-estimation.pdf.