Linear Correlation

This function computed the linear correlation between two vectors or a correlation matrix for an input matrix.

The following methods to compute linear correlations are implemented in this function:

lin.cor(x, y = NULL, method = "pearson", test.na = FALSE)

Arguments

x
a numeric vector, matrix, or data.frame.

y
a numeric vector that should be correlated with x.

method
the method to compute the linear correlation between x and y.

test.na
a boolean value indicating whether input data should be checked for NA values.

Details

method = "pearson" : Pearson's correlation coefficient (centred).

method = "pearson2" : Pearson's uncentred correlation coefficient.

method = "sq_pearson" . Squared Pearson's correlation coefficient.

method = "kendall" : Kendall's correlation coefficient.

method = "spearman" : Spearman's correlation coefficient.

Further Details:

Pearson's correlation coefficient (centred) :

Author

Hajk-Georg Drost