Distance Diversity between Probability Density Functions

This function computes all distance values between two probability density functions that are available in getDistMethods and returns a vector storing the corresponding distance measures. This vector is named distance diversity vector.

dist.diversity(x, p, test.na = FALSE, unit = "log2")

Arguments

x
a numeric data.frame or matrix (storing probability vectors) or a numeric data.frame or matrix storing counts (if est.prob is specified).

p
power of the Minkowski distance.

test.na
a boolean value indicating whether input vectors should be tested for NA values. Faster computations if test.na = FALSE.

unit
a character string specifying the logarithm unit that should be used to compute distances that depend on log computations. Options are:

unit = "log"

unit = "log2"

unit = "log10"

Examples

dist.diversity(rbind(1:10/sum(1:10), 20:29/sum(20:29)), p = 2, unit = "log2")

#> Metric: 'euclidean' using unit: 'log2'.

#> Metric: 'manhattan' using unit: 'log2'.

#> Metric: 'minkowski' using unit: 'log2'.

#> Metric: 'chebyshev' using unit: 'log2'.

#> Metric: 'sorensen' using unit: 'log2'.

#> Metric: 'gower' using unit: 'log2'.

#> Metric: 'soergel' using unit: 'log2'.

#> Metric: 'kulczynski_d' using unit: 'log2'.

#> Metric: 'canberra' using unit: 'log2'.

#> Metric: 'lorentzian' using unit: 'log2'.

#> Metric: 'intersection' using unit: 'log2'.

#> Metric: 'non-intersection' using unit: 'log2'.

#> Metric: 'wavehedges' using unit: 'log2'.

#> Metric: 'czekanowski' using unit: 'log2'.

#> Metric: 'motyka' using unit: 'log2'.

#> Metric: 'kulczynski_s' using unit: 'log2'.

#> Metric: 'tanimoto' using unit: 'log2'.

#> Metric: 'ruzicka' using unit: 'log2'.

#> Metric: 'inner_product' using unit: 'log2'.

#> Metric: 'harmonic_mean' using unit: 'log2'.

#> Metric: 'cosine' using unit: 'log2'.

#> Metric: 'hassebrook' using unit: 'log2'.

#> Metric: 'jaccard' using unit: 'log2'.

#> Metric: 'dice' using unit: 'log2'.

#> Metric: 'fidelity' using unit: 'log2'.

#> Metric: 'bhattacharyya' using unit: 'log2'.

#> Metric: 'hellinger' using unit: 'log2'.

#> Metric: 'matusita' using unit: 'log2'.

#> Metric: 'squared_chord' using unit: 'log2'.

#> Metric: 'squared_euclidean' using unit: 'log2'.

#> Metric: 'pearson' using unit: 'log2'.

#> Metric: 'neyman' using unit: 'log2'.

#> Metric: 'squared_chi' using unit: 'log2'.

#> Metric: 'prob_symm' using unit: 'log2'.

#> Metric: 'divergence' using unit: 'log2'.

#> Metric: 'clark' using unit: 'log2'.

#> Metric: 'additive_symm' using unit: 'log2'.

#> Metric: 'kullback-leibler' using unit: 'log2'.

#> Metric: 'jeffreys' using unit: 'log2'.

#> Metric: 'k_divergence' using unit: 'log2'.

#> Metric: 'topsoe' using unit: 'log2'.

#> Metric: 'jensen-shannon' using unit: 'log2'.

#> Metric: 'jensen_difference' using unit: 'log2'.

#> Metric: 'taneja' using unit: 'log2'.

#> Metric: 'kumar-johnson' using unit: 'log2'.

#> Metric: 'avg' using unit: 'log2'.

#>         euclidean         manhattan         minkowski         chebyshev 
#>        0.12807130        0.35250464        0.12807130        0.06345083 
#>          sorensen             gower           soergel      kulczynski_d 
#>        0.17625232        0.03525046        0.29968454        0.42792793 
#>          canberra        lorentzian      intersection  non-intersection 
#>        2.09927095        0.49712136        0.82374768        0.17625232 
#>        wavehedges       czekanowski            motyka      kulczynski_s 
#>        3.16657887        0.17625232        0.58812616        2.33684211 
#>          tanimoto           ruzicka     inner_product     harmonic_mean 
#>        0.29968454        0.70031546        0.10612245        0.94948528 
#>            cosine        hassebrook           jaccard              dice 
#>        0.93427641        0.86613103        0.13386897        0.07173611 
#>          fidelity     bhattacharyya         hellinger          matusita 
#>        0.97312397        0.03930448        0.32787819        0.23184489 
#>     squared_chord squared_euclidean           pearson            neyman 
#>        0.05375205        0.01640226        0.16814418        0.36742465 
#>       squared_chi         prob_symm        divergence             clark 
#>        0.10102943        0.20205886        1.49843905        0.86557468 
#>     additive_symm  kullback-leibler          jeffreys      k_divergence 
#>        0.53556883        0.13926288        0.31761069        0.04216273 
#>            topsoe    jensen-shannon jensen_difference            taneja 
#>        0.07585498        0.03792749        0.03792749        0.04147518 
#>     kumar-johnson               avg 
#>        0.62779644        0.20797774